Refactor CoverEditor component
Splits the CoverEditor component into smaller, more manageable files to improve code organization and maintainability.
